Create polling util
