+++ packages/cli/lib/polling.js
@@ -1,0 +1,22 @@
+const { POLLING_DELAY } = require('./constants');
+
+const poll = (callback, accountId, taskId) => {
+  return new Promise((resolve, reject) => {
+    const pollInterval = setInterval(async () => {
+      const pollResp = await callback(accountId, taskId);
+      const { status } = pollResp;
+
+      if (status === 'SUCCESS') {
+        clearInterval(pollInterval);
+        resolve(pollResp);
+      } else if (status === 'ERROR') {
+        clearInterval(pollInterval);
+        reject(pollResp);
+      }
+    }, POLLING_DELAY);
+  });
+};
+
+module.exports = {
+  poll,
+};
